requirements.txt
Flask
firebase-admin
google-generativeai
python-dotenv